+++ src/control/songLibLayout.fxml
@@ -1,0 +1,121 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+<?import javafx.scene.text.*?>
+
+<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" 
+		minWidth="-Infinity" prefHeight="421.0" prefWidth="469.0" 
+		xmlns="http://javafx.com/javafx/15.0.1" 
+		xmlns:fx="http://javafx.com/fxml/1"
+		fx:controller="control.SongController">
+   <children>
+      <ListView layoutX="6.0" layoutY="6.0" prefHeight="407.0" prefWidth="235.0" />
+      <Text layoutX="301.0" layoutY="31.0" strokeType="OUTSIDE" strokeWidth="0.0" 
+      	text="Song Details">
+         <font>
+            <Font name="System Bold" size="18.0" />
+         </font>
+      </Text>
+      <GridPane layoutX="239.0" layoutY="42.0" prefHeight="123.0" prefWidth="219.0">
+        <columnConstraints>
+          <ColumnConstraints hgrow="SOMETIMES" maxWidth="95.0" minWidth="10.0" prefWidth="60.0" />
+          <ColumnConstraints hgrow="SOMETIMES" maxWidth="167.0" minWidth="10.0" prefWidth="159.0" />
+        </columnConstraints>
+        <rowConstraints>
+          <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+          <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+          <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+            <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+        </rowConstraints>
+         <children>
+            <Text strokeType="OUTSIDE" strokeWidth="0.0" text="Title" textAlignment="CENTER" 
+            	wrappingWidth="58.13671875" />
+            <Text strokeType="OUTSIDE" strokeWidth="0.0" text="Artist" textAlignment="CENTER" 
+            	wrappingWidth="60.13671875" GridPane.rowIndex="1" />
+            <Text strokeType="OUTSIDE" strokeWidth="0.0" text="Album" textAlignment="CENTER" 
+            	wrappingWidth="59.13671875" GridPane.rowIndex="2" />
+            <Text strokeType="OUTSIDE" strokeWidth="0.0" text="Year" textAlignment="CENTER" 
+            	wrappingWidth="58.13671875" GridPane.rowIndex="3" />
+            <Label prefHeight="25.0" prefWidth="159.0" textAlignment="CENTER" 
+            	GridPane.columnIndex="1">
+               <font>
+                  <Font size="13.0" />
+               </font>
+            </Label>
+            <Label prefHeight="25.0" prefWidth="159.0" textAlignment="CENTER" 
+            	GridPane.columnIndex="1" GridPane.rowIndex="1">
+               <font>
+                  <Font size="13.0" />
+               </font>
+            </Label>
+            <Label prefHeight="23.0" prefWidth="159.0" textAlignment="CENTER" 
+            	GridPane.columnIndex="1" GridPane.rowIndex="2">
+               <font>
+                  <Font size="13.0" />
+               </font>
+            </Label>
+            <Label prefHeight="24.0" prefWidth="159.0" GridPane.columnIndex="1" 
+            	GridPane.rowIndex="3">
+               <font>
+                  <Font size="13.0" />
+               </font>
+            </Label>
+         </children>
+      </GridPane>
+      <Text layoutX="298.0" layoutY="198.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Song Controls">
+         <font>
+            <Font name="System Bold" size="18.0" />
+         </font>
+      </Text>
+      <GridPane layoutX="239.0" layoutY="216.0" prefHeight="123.0" prefWidth="219.0">
+         <columnConstraints>
+            <ColumnConstraints hgrow="SOMETIMES" maxWidth="95.0" minWidth="10.0" prefWidth="60.0" />
+            <ColumnConstraints hgrow="SOMETIMES" maxWidth="167.0" minWidth="10.0" prefWidth="159.0" />
+         </columnConstraints>
+         <rowConstraints>
+            <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+            <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+            <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+            <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+         </rowConstraints>
+         <children>
+            <Text strokeType="OUTSIDE" strokeWidth="0.0" text="Title" textAlignment="CENTER" 
+            	wrappingWidth="58.13671875" />
+            <Text strokeType="OUTSIDE" strokeWidth="0.0" text="Artist" textAlignment="CENTER" 
+            	wrappingWidth="60.13671875" GridPane.rowIndex="1" />
+            <Text strokeType="OUTSIDE" strokeWidth="0.0" text="Album" textAlignment="CENTER" 
+            	wrappingWidth="59.13671875" GridPane.rowIndex="2" />
+            <Text strokeType="OUTSIDE" strokeWidth="0.0" text="Year" textAlignment="CENTER" 
+            	wrappingWidth="58.13671875" GridPane.rowIndex="3" />
+            <TextField GridPane.columnIndex="1" />
+            <TextField GridPane.columnIndex="1" GridPane.rowIndex="1" />
+            <TextField GridPane.columnIndex="1" GridPane.rowIndex="2" />
+            <TextField GridPane.columnIndex="1" GridPane.rowIndex="3" />
+         </children>
+      </GridPane>
+      <Button fx:id="addB" layoutX="286.0" layoutY="346.0" mnemonicParsing="false" prefHeight="30.0" 
+      	prefWidth="63.0" text="Add">
+         <font>
+            <Font name="System Bold" size="14.0" />
+         </font>
+      </Button>
+      <Button fx:id="editB" layoutX="375.0" layoutY="346.0" mnemonicParsing="false" prefHeight="30.0" 
+      	prefWidth="63.0" text="Edit">
+         <font>
+            <Font name="System Bold" size="14.0" />
+         </font>
+      </Button>
+      <Button fx:id="deleteB" layoutX="286.0" layoutY="383.0" mnemonicParsing="false" text="Delete">
+         <font>
+            <Font name="System Bold" size="14.0" />
+         </font>
+      </Button>
+      <Button fx:id="saveB" layoutX="375.0" layoutY="383.0" mnemonicParsing="false" prefHeight="30.0" 
+      	prefWidth="63.0" text="Save">
+         <font>
+            <Font name="System Bold" size="14.0" />
+         </font>
+      </Button>
+   </children>
+</AnchorPane>
